+++ web/static/react/assets/BackupPage-D2mqRXwO.js
@@ -1,1 +1,0 @@
-import{t as e}from"./vendor-react-B-Deo4S-.js";import{t}from"./ImpostazioniPage-Ct5_M0RO.js";var n=e();function r(){return(0,n.jsx)(t,{})}export{r as BackupPage};