+++ web/static/react/assets/SkillRunDetailPage-DWYQnLAj.js
@@ -1,0 +1,1 @@
+import{t as e}from"./vendor-react-B-Deo4S-.js";import{t}from"./LegalSkillsReviewPage-CRkIuThX.js";var n=e();function r(){return(0,n.jsx)(t,{})}export{r as SkillRunDetailPage};